+++ backend/src/main/java/com/dev258/rota365/usuario/entity/Usuario.java
@@ -25,8 +25,8 @@ public class Usuario {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_usuario")
-    private Long idUsuario;
-
+    private Integer idUsuario;
+    
     @Column(name = "nome", nullable = false, length = 100)
     private String nome;
